+++ src/main/java/org/c4marathon/assignment/repository/mainAccount/MainAccountQueryRepository.java
@@ -1,4 +1,4 @@
-package org.c4marathon.assignment.repository;
+package org.c4marathon.assignment.repository.mainAccount;
 
 public interface MainAccountQueryRepository {
 	int conditionalFastCharge(Long id, Long amount, Long minRequiredBalance, Long limit);
